content_marketing_agent/requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.0
markdown>=3.5.0
pymongo>=4.8.0

# LangChain stack
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=0.3.0
langchain-pinecone>=0.1.0
langchain-huggingface>=0.1.0
langgraph>=0.1.0
langchain-core>=0.3.0
pinecone-client>=5.0.0

# HTTP client for Perplexity
requests>=2.31.0
